requirements.txt
anycache==2.0.7
asgiref==3.2.3
dill==0.3.1.1
Django==3.0.4
djangorestframework==3.11.0
filelock==3.0.12
numpy==1.18.1
pandas==1.0.1
pkg-resources==0.0.0
psycopg2-binary==2.8.4
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
xlrd==1.2.0
